--- client/src/components/Footer/Footer.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import './styles.scss';
-
-const Footer = () => {
-  return (
-    <div className="footer">
-      <div className="footer-content">
-        <span className="username">@Cookinity</span>
-      </div>
-    </div>
-  );
-};
-
-export default Footer;
