--- client/src/components/Loader/styles.scss
@@ -1,10 +1,0 @@
-.loader-container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  /* align-items: center; */
-}
-
-.loader .loader-content {
-}
