--- client/src/hoc/requireAdmin.js
@@ -1,32 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-export default (ChildComponent) => {
-  class ComposedComponent extends Component {
-    // Our component just got rendered
-    componentDidMount() {
-      this.shouldNavigateAway();
-    }
-
-    // Our component just got updated
-    componentDidUpdate() {
-      this.shouldNavigateAway();
-    }
-
-    shouldNavigateAway() {
-      if (!(this.props.auth.isAuthenticated && this.props.auth.me?.role === 'ADMIN')) {
-        this.props.history.push('/');
-      }
-    }
-
-    render() {
-      return <ChildComponent {...this.props} />;
-    }
-  }
-
-  function mapStateToProps(state) {
-    return { auth: state.auth };
-  }
-
-  return connect(mapStateToProps)(ComposedComponent);
-};
