--- client/src/hoc/requireAuth.js
@@ -1,34 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-
-export default (ChildComponent) => {
-  class ComposedComponent extends Component {
-    // Our component just got rendered
-    componentDidMount() {
-      this.shouldNavigateAway();
-    }
-
-    // Our component just got updated
-    componentDidUpdate() {
-      this.shouldNavigateAway();
-    }
-
-    //token ima auth nema, prepisi ga sa func component i useefect
-    shouldNavigateAway() {
-      if (!this.props.auth.isAuthenticated) {
-        console.log('history push');
-        this.props.history.push('/');
-      }
-    }
-
-    render() {
-      return <ChildComponent {...this.props} />;
-    }
-  }
-
-  function mapStateToProps(state) {
-    return { auth: state.auth };
-  }
-
-  return connect(mapStateToProps)(ComposedComponent);
-};
