--- client/src/pages/Home/Home.js
@@ -1,42 +1,0 @@
-import React from 'react';
-import { compose } from 'redux';
-import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
-
-import Layout from '../../layout/Layout';
-
-const Home = ({ auth }) => {
-  return (
-    <Layout>
-      <div>
-        <h1>Home page</h1>
-        {!auth.isAuthenticated ? (
-          <div>
-            <p>
-              Welcome guest!{' '}
-              <Link className="bold" to="/login">
-                Log in
-              </Link>{' '}
-              or{' '}
-              <Link className="bold" to="/register">
-                Register
-              </Link>
-            </p>
-          </div>
-        ) : (
-          <>
-            <p>
-              Welcome <span className="name">{auth.me.name}</span>!
-            </p>
-          </>
-        )}
-      </div>
-    </Layout>
-  );
-};
-
-const mapStateToProps = (state) => ({
-  auth: state.auth,
-});
-
-export default compose(connect(mapStateToProps))(Home);
