--- client/src/pages/NotFound/styles.scss
@@ -1,3 +1,0 @@
-.not-found-page {
-  text-align: center;
-}
