--- client/src/pages/Register/styles.scss
@@ -1,14 +1,0 @@
-.register {
-  display: flex;
-  justify-content: center;
-}
-
-.register .container {
-  max-width: 400px;
-  flex: 1;
-}
-
-.register .submit {
-  background-color: gray;
-  color: white;
-}
