+++ client/src/store/features/users/usersReducer.js
@@ -1,4 +1,4 @@
-import { GET_USERS_LOADING, GET_USERS_SUCCESS, GET_USERS_FAIL } from '../../types';
+import { GET_USERS_LOADING, GET_USERS_SUCCESS, GET_USERS_FAIL } from '../../actionTypes';
 
 const initialState = {
   users: [],
