--- client/src/store/types.js
@@ -1,32 +1,0 @@
-export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-
-export const LOGIN_WITH_EMAIL_LOADING = 'LOGIN_WITH_EMAIL_LOADING';
-export const LOGIN_WITH_EMAIL_SUCCESS = 'LOGIN_WITH_EMAIL_SUCCESS';
-export const LOGIN_WITH_EMAIL_FAIL = 'LOGIN_WITH_EMAIL_FAIL';
-
-export const REGISTER_WITH_EMAIL_LOADING = 'REGISTER_WITH_EMAIL_LOADING';
-export const REGISTER_WITH_EMAIL_SUCCESS = 'REGISTER_WITH_EMAIL_SUCCESS';
-export const REGISTER_WITH_EMAIL_FAIL = 'REGISTER_WITH_EMAIL_FAIL';
-
-export const ME_LOADING = 'ME_LOADING';
-export const ME_SUCCESS = 'ME_SUCCESS';
-export const ME_FAIL = 'ME_FAIL';
-
-export const GET_FEATURE = 'GET_FEATURE';
-export const GET_PROFILE = 'GET_PROFILE';
-
-export const GET_USERS_LOADING = 'GET_USERS_LOADING';
-export const GET_USERS_SUCCESS = 'GET_USERS_SUCCESS';
-export const GET_USERS_FAIL = 'GET_USERS_FAIL';
-
-export const GET_PROFILE_LOADING = 'GET_PROFILE_LOADING';
-export const GET_PROFILE_SUCCESS = 'GET_PROFILE_SUCCESS';
-export const GET_PROFILE_FAIL = 'GET_PROFILE_FAIL';
-
-export const EDIT_USER_LOADING = 'EDIT_USER_LOADING';
-export const EDIT_USER_SUCCESS = 'EDIT_USER_SUCCESS';
-export const EDIT_USER_FAIL = 'EDIT_USER_FAIL';
-
-export const DELETE_USER_LOADING = 'DELETE_USER_LOADING';
-export const DELETE_USER_SUCCESS = 'DELETE_USER_SUCCESS';
-export const DELETE_USER_FAIL = 'DELETE_USER_FAIL';
